+++ src/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/[lessonSlug]/Components/DownloadSuccessHeader/DownloadSuccessHeader.tsx
@@ -8,20 +8,38 @@ import {
   OakSpan,
   OakTertiaryInvertedButton,
 } from "@oaknational/oak-components";
+import { useOakConsent } from "@oaknational/oak-consent-client";
 
-import { Header } from "@/components/TeacherComponents/Header/Header";
+import {
+  Header,
+  HeaderProps,
+} from "@/components/TeacherComponents/Header/Header";
 import { getCloudinaryImageUrl } from "@/utils/getCloudinaryImageUrl";
 import { resolveOakHref } from "@/common-lib/urls";
+import { ServicePolicyMap } from "@/browser-lib/cookie-consent/ServicePolicyMap";
 
 const DOWNLOAD_SUCCESS_IMG_URL =
   "v1777386544/svg-illustrations/download-confirmation-Illustration_z1sczk.svg";
 
-export function DownloadSuccessHeader({ href }: Readonly<{ href: string }>) {
+type DownloadSuccessHeaderProps = {
+  href?: string;
+  onBackClick?: () => void;
+  backgroundColorLevel?: HeaderProps["backgroundColorLevel"];
+  returnTo: "lesson" | "downloads";
+};
+
+export function DownloadSuccessHeader(
+  props: Readonly<DownloadSuccessHeaderProps>,
+) {
+  /** We only show the help message if the user has consented to the Gleap cookie */
+  const { getConsent } = useOakConsent();
+  const cookiesNotAccepted = getConsent(ServicePolicyMap.GLEAP) === "denied";
+
   return (
     <Header
       layoutVariant="large"
       useSubduedBackground
-      headerSlot={<BackLink href={href} />}
+      headerSlot={<BackLinkButton {...props} />}
       heading="Thanks for downloading!"
       summary={
         <OakFlex $flexDirection="column" $gap={"spacing-24"}>
@@ -29,16 +47,20 @@ export function DownloadSuccessHeader({ href }: Readonly<{ href: string }>) {
             We hope you find the resources useful. Click the question mark in
             the bottom-right corner to share your feedback.{" "}
           </OakP>
-          <InstallFontsInstructions />
+          <InstallFontsInstructions showHelpMessage={!cookiesNotAccepted} />
         </OakFlex>
       }
-      backgroundColorLevel={1}
+      backgroundColorLevel={props.backgroundColorLevel}
       heroImage={getCloudinaryImageUrl(DOWNLOAD_SUCCESS_IMG_URL)}
     />
   );
 }
 
-function InstallFontsInstructions() {
+function InstallFontsInstructions({
+  showHelpMessage,
+}: Readonly<{
+  showHelpMessage: boolean;
+}>) {
   return (
     <OakFlex $gap="spacing-8" $alignItems={"center"}>
       <OakIcon iconWidth="spacing-20" iconName="info" />
@@ -51,28 +73,31 @@ function InstallFontsInstructions() {
             "install the Google Fonts 'Lexend' and 'Kalam' (opens in a new tab)"
           }
         >
-          install the Google Fonts ‘Lexend’ and ‘Kalam’
+          install the Google Fonts ‘Lexend’ and ‘Kalam’.{" "}
         </OakLink>
-        <OakSpan>
-          . Click the question mark in the bottom-right of the page if you need
-          extra help with this.
-        </OakSpan>
+        {showHelpMessage && (
+          <OakSpan>
+            Click the question mark in the bottom-right of the page if you need
+            extra help with this.
+          </OakSpan>
+        )}
       </OakP>
     </OakFlex>
   );
 }
 
-function BackLink({ href }: Readonly<{ href: string }>) {
+function BackLinkButton(props: Readonly<DownloadSuccessHeaderProps>) {
   return (
     <OakBox>
       <OakTertiaryInvertedButton
-        element={Link}
-        href={href}
+        element={props.href ? Link : "button"}
+        href={props.href ?? undefined}
         aria-label={"Back to lesson"}
         iconName={"arrow-left"}
         isTrailingIcon={false}
+        onClick={props.onBackClick}
       >
-        Back to lesson
+        Back to {props.returnTo}
       </OakTertiaryInvertedButton>
     </OakBox>
   );