+++ src/node-lib/curriculum-api-2023/queries/teachersUnitOverview/helpers.ts
@@ -323,7 +323,9 @@ export const getProgrammeToggles = (
           programme.programme_fields.subject_slug ===
             currentProgramme.programme_fields.subject_slug &&
           programme.programme_fields.examboard_slug ===
-            currentProgramme.programme_fields.examboard_slug,
+            currentProgramme.programme_fields.examboard_slug &&
+          programme.programme_fields.pathway_slug ===
+            currentProgramme.programme_fields.pathway_slug,
       )
       .filter(hasTierFields)
       .toSorted((a, b) => sortTiers(a.programme_fields, b.programme_fields))
@@ -340,7 +342,9 @@ export const getProgrammeToggles = (
         programme.programme_fields.examboard_slug ===
           currentProgramme.programme_fields.examboard_slug &&
         programme.programme_fields.tier_slug ===
-          currentProgramme.programme_fields.tier_slug,
+          currentProgramme.programme_fields.tier_slug &&
+        programme.programme_fields.pathway_slug ===
+          currentProgramme.programme_fields.pathway_slug,
     )
     .toSorted((a, b) =>
       sortChildSubjects(a.programme_fields, b.programme_fields),